+++ src/lib/layeredCanvas/tools/draw/drawFilmStack.ts
@@ -22,6 +22,9 @@ export function drawFilmStack(ctx: CanvasRenderingContext2D, filmStack: FilmStac
     ctx.rotate(-film.rotation * Math.PI / 180);
     ctx.scale(scale * film.reverse[0], scale * film.reverse[1]);
     ctx.globalAlpha *= (film.opacity ?? 1.0);
+    if (film.blendMode && film.blendMode !== 'source-over') {
+      ctx.globalCompositeOperation = film.blendMode;
+    }
 
     if (film.content.kind === 'media') {
       let media = film.content.media;